+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.0.4
 - Add StrSubstitutor class.
+- Add StrBuilder class.
+- Update readme.
+- Add user guide.
 
 ## 0.0.3
 - Add StrLookup class.